Refactor controller

+++ src/osspec.h
@@ -3,6 +3,6 @@
 
 void entropy(uint8_t *buf, size_t len);
 
-void get_time(uint32_t *sec, uint32_t *usec);
+void get_time(uint32_t *sec, uint16_t *usec);
 
 #endif /* _OSSPEC_H */
